flask_rest_app/requirements.txt
alabaster==0.7.11
aniso8601==3.0.2
Babel==2.6.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
Flask==1.0.2
Flask-HTTPAuth==3.2.4
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
gunicorn==19.9.0
idna==2.7
imagesize==1.0.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
packaging==17.1
Pillow==5.2.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2018.5
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
